[TASK] Move TYPO3 modal logic into new helper `Typo3ModalDialog`
