+++ packages/screenshots/Classes/Runner/codeception.yml
@@ -85,6 +85,7 @@ modules:
     - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3Draw
     - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3FileSystem
     - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3FileTree
+    - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3ModalDialog
     - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3Navigation
     - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3PageTree
     - TYPO3\CMS\Screenshots\Runner\Codeception\Support\Helper\Typo3Screenshots:
